refactor: move funtions to own file
